backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.3.1
Flask-Cors==3.0.10
python-dotenv==0.19.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
bcrypt==3.2.0
email-validator==1.1.3
marshmallow==3.13.0
